migration default tinyInteger.

+++ database/migrations/2019_09_19_035651_create_formations_table.php
@@ -23,7 +23,7 @@ class CreateFormationsTable extends Migration
             $table->string('code', 10);
             $table->integer('length');
             $table->string('label');
-            $table->integer('state')->default(1);
+            $table->tinyInteger('state')->default(1);
             $table->text('description');
             $table->timestamps();
             $table->integer('former_id');
